move dotweb/state to dotweb/core/state
