+++ core/state.go
@@ -1,4 +1,4 @@
-package dotweb
+package core
 
 import (
 	"sync/atomic"
